requirements.txt
# Core FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud services
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-auth==2.23.4

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling and uploads
python-multipart==0.0.6
pillow==10.1.0
aiofiles==23.2.1

# QR code generation
qrcode[pil]==7.4.2

# Environment and configuration
python-dotenv==1.0.0

# Email validation
email-validator==2.1.0

# HTTP client for external APIs
httpx==0.25.2

# Template engine (for email templates if needed)
jinja2==3.1.2

# Health check support
requests==2.31.0